Rename classes to work with CSS module
