+++ client/utils/formatClasses.js
